refactor(runtime): split timer + copying-GC test tails into sibling modules
The 2000-line file gate (scripts/check_file_size.sh) failed the lint job:
timer.rs was 1998 lines on main -- two under the cap -- so the drain
partition, the expired-batch ordering, and their tests pushed it to 2163,
and the new Map pointer-index GC tests took gc/tests/copying.rs to 2040.

Mechanical cut, no behavior change:

- timer.rs: the whole trailing #[cfg(test)] region (test-only seed/snapshot
  helpers + the drain/ordering unit tests) moves to timer/tests.rs, next to
  the existing timer/ref_states.rs. The pub(crate) helpers are re-exported
  from timer.rs, so out-of-module call sites (crate::timer::test_seed_*,
  used by the gc root-scanner tests) resolve unchanged. 2163 -> 1913.
- gc/tests/copying.rs: the two Map pointer-key index tests move to
  copying/map_pointer_index.rs, alongside the existing copying/ siblings.
  2040 -> 1940.

map.rs stays over the gate but is already allowlisted.

cargo test -p perry-runtime: 1239 passed, 0 failed (unchanged). fmt,
file-size, GC store-site inventory and address-classification gates all
pass locally.
